Tests assembly and hierarchy refactor
Remaniement du dossier test avec référence (Assembly Definition) et tests pour le play mode et edit mode
