+++ Assets/Input Actions.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: cd23dfe00bb2a7c44af1b9c7fb70a46b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
