+++ Assets/Scripts/BeachObjects.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c5ef3934735567d4c89140f61c5bb9ad
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
