+++ Assets/Scripts/BeachObjects/BeachObjects.asmdef
@@ -1,0 +1,3 @@
+{
+	"name": "BeachObjects"
+}
