+++ Assets/Scripts/BeachObjects/BeachObjects.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 0e554704d47a99a46b6f95f457541c6c
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
