+++ Assets/Scripts/BeachObjects/Platform.cs
@@ -1,71 +1,9 @@
 using System.Collections;
 using System.Collections.Generic;
-using NUnit.Framework;
 using UnityEngine;
-using UnityEngine.TestTools;
 
-public class TestPlatform
+namespace BeachObjects
 {
-    private Platform _platform;
-
-    [Test]
-    public void _0_Sets_Platform_to_Max_size()
-    {
-        float maxSize = 80f;
-        _platform = new Platform();
-
-        _platform.SetPlatformSize(maxSize);
-
-        Assert.AreEqual(maxSize, _platform.GetPlatformSize());
-    }
-
-    [Test]
-    public void _1_Sets_Platform_to_Min_size()
-    {
-        float minSize = 45f;
-        _platform = new Platform();
-
-        _platform.SetPlatformSize(minSize);
-
-        Assert.AreEqual(minSize, _platform.GetPlatformSize());
-    }
-
-    [Test]
-    public void _2_Sets_Number_of_Swarm_to_Max()
-    {
-        int max = 5;
-        _platform = new Platform(max);
-
-        Assert.AreEqual(max, _platform.GetNbSwarm());
-    }
-
-    [Test]
-    public void _3_Sets_Number_of_Swarm_to_Min()
-    {
-        int min = 1;
-        _platform = new Platform(min);
-
-        Assert.AreEqual(min, _platform.GetNbSwarm());
-    }
-
-    [Test]
-    public void _4_Sets_Distance_From_Cost_to_Max()
-    {
-        float maxDistance = 490;
-        _platform = new Platform(1, maxDistance);
-
-        Assert.AreEqual(maxDistance, _platform.GetDistanceFromCost());
-    }
-
-    [Test]
-    public void _5_Sets_Distance_From_Cost_to_Max()
-    {
-        float minDistance = 150;
-        _platform = new Platform(1, minDistance);
-
-        Assert.AreEqual(minDistance, _platform.GetDistanceFromCost());
-    }
-
     public class Platform
     {
         private BoxCollider _surface;
@@ -90,7 +28,7 @@ public class TestPlatform
                 swarmNumber = 1;
             else if (swarmNumber > 5)
                 swarmNumber = 5;
-            
+
             _surface = new GameObject().AddComponent<BoxCollider>();
             _surface.transform.localScale = new Vector3(45.0f, 1.0f, 45.0f);
             _swarmNumber = swarmNumber;
@@ -110,7 +48,7 @@ public class TestPlatform
                 swarmNumber = 1;
             else if (swarmNumber > 10)
                 swarmNumber = 5;
-            
+
             _surface = new GameObject().AddComponent<BoxCollider>();
             _surface.transform.localScale = new Vector3(45.0f, 1.0f, 45.0f);
             _swarmNumber = swarmNumber;
@@ -161,24 +99,6 @@ public class TestPlatform
             Vector3 scale = _surface.transform.localScale;
 
             _surface.transform.localScale = new Vector3(scale.x, scale.y, size);
-        }
-    }
-
-    public class Swarm
-    {
-        private int _id;
-        private List<object> _droneList;
-
-        public Swarm(int id)
-        {
-            _id = id;
-            _droneList = new List<object>();
-
-            for (int i = 0; i < 4; i++)
-            {
-                object newDrone = new object();
-                _droneList.Add(newDrone);
-            }
         }
     }
 }
