+++ Assets/Scripts/BeachObjects/Platform.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f14b16552f6521749b64025bf0f50ec3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
