+++ Assets/Scripts/Drone.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9d6527dd67e7ee543a458c7705318b9d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
