+++ Assets/Scripts/Drone/BaseRigibody.cs
@@ -1,0 +1,30 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace Drone
+{
+    [RequireComponent(typeof(Rigidbody))]
+    public class BaseRigibody : MonoBehaviour
+    {
+        #region Variables
+        
+        #endregion
+
+        #region Main Methods
+        void Start()
+        {
+            
+        }
+
+        void Update()
+        {
+            
+        }
+        #endregion
+
+        #region Custom Methods
+        
+        #endregion
+    }
+}
