+++ Assets/Scripts/Drone/BaseRigibody.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9b8345c4bdb8d4d4f9acf275b440a1e1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
