+++ Assets/Scripts/Drone/Drone.asmdef
@@ -1,0 +1,3 @@
+{
+	"name": "Drone"
+}
