+++ Assets/Scripts/Drone/Drone.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 853c267b6374df74c96a63aca63645b6
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
