+++ Assets/Scripts/Main.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 696a49453b596a342b29e1d83f0d158c
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
