--- Assets/Scripts/Tests/Tests.asmdef
@@ -1,9 +1,0 @@
-{
-    "name": "Tests",
-    "optionalUnityReferences": [
-        "TestAssemblies"
-    ],
-    "includePlatforms": [
-        "Editor"
-    ]
-}
