+++ Assets/Tests/EditorMode/DroneTests.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e9e407df01f92ab41a52b516cb1665bd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
