+++ Assets/Tests/EditorMode/InputTests.cs
@@ -1,0 +1,14 @@
+using System.Collections;
+using System.Collections.Generic;
+using NUnit.Framework;
+using UnityEngine;
+using UnityEngine.TestTools;
+
+public class InputTests
+{
+    [Test]
+    public void _0_Go_Up ()
+    {
+        
+    }
+}
