+++ Assets/Tests/EditorMode/InputTests.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e65daa7eed7cfe84194ffe10400f891d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
