+++ Assets/Tests/PlayMode/NewTestScript.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 61f4e5a24a787264dbd3e325a431a7b5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
