+++ ProjectSettings/ProjectSettings.asset
@@ -152,7 +152,8 @@ PlayerSettings:
   resolutionScalingMode: 0
   androidSupportedAspectRatio: 1
   androidMaxAspectRatio: 2.1
-  applicationIdentifier: {}
+  applicationIdentifier:
+    Standalone: com.DefaultCompany.Surveillancedeplage
   buildNumber:
     Standalone: 0
     iPhone: 0
